# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Radware Vulnerable E-commerce API",
    "version": "1.2.0",
    "description": "An intentionally vulnerable e-commerce API designed to demonstrate business logic attacks, focusing on path and query parameters. This version includes Protected Entity Mode which blocks destructive actions on certain demo entities.\n"
  },
...
  "paths": {
    "/api/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create a new user",
        "description": "Creates a new user. This endpoint is distinct from /api/auth/register and might have different handling or intended use in a larger system, but here it functions similarly. It does not require authentication.",
        "operationId": "createUserFromRootPath",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Username for the new user"
          },
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "email"
            },
            "description": "Email for the new user"
          },
          {
            "name": "password",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "password"
            },
            "description": "Password for the new user"
          }
        ],
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (e.g., username/email already exists)"
          }
        }
      },
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List all users",
        "description": "Retrieve a list of all users in the system. This endpoint demonstrates a BOLA/BFLA vulnerability by allowing any authenticated user to list all users.",
        "operationId": "listAllUsers",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Bearer token is missing or invalid"
          }
        }
      }
    },
    "/api/users/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user by user ID (BOLA Target)",
        "description": "Retrieves user details. BOLA vulnerability: Any user (even unauthenticated for this specific endpoint as implemented in user_router) can view any other user's details if the user_id is known.",
        "operationId": "getUserByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the user to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update user by user ID (BOLA & Parameter Pollution Target)",
        "description": "Updates user details. BOLA: Any user can attempt to update another user. Parameter Pollution: `is_admin` can be passed to escalate privileges. Protected users have restrictions on username/email changes.",
        "operationId": "updateUserByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the user to update"
          },
          {
            "name": "username",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "New username for the user"
          },
          {
            "name": "email",
            "in": "query",
            "required": false,
            "allowReserved": true,
            "schema": {
              "type": "string",
              "format": "email"
            },
            "description": "New email for the user"
          },
          {
            "name": "is_admin",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            },
            "description": "Set user as admin (Parameter Pollution vulnerability)"
          }
        ],
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Action Forbidden: The user is protected for demo purposes (e.g., trying to change username/email of a protected user)."
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete user by user ID (BFLA/BOLA Target)",
        "description": "Deletes a user. BFLA: No admin check is performed. BOLA: Any user can attempt to delete any other user. Protected users cannot be deleted.",
        "operationId": "deleteUserByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the user to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Action Forbidden: The user is protected for demo purposes and cannot be deleted."
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Log in a user",
        "description": "User login using query parameters to obtain a JWT.",
        "operationId": "loginUser",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Username"
          },
          {
            "name": "password",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "password"
            },
            "description": "Password",
            "allowReserved": true,
            "allowEmptyValue": false
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful, JWT returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "required": [
          "user_id",
          "username",
          "email"
        ],
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the user",
            "readOnly": true
          },
          "username": {
            "type": "string",
            "description": "Username of the user (unique)"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the user (unique)"
          },
          "password_hash": {
            "type": "string",
            "description": "Hashed password of the user",
            "readOnly": true
          },
          "is_admin": {
            "type": "boolean",
            "default": false,
            "description": "Flag indicating if the user is an administrator"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of user creation",
            "readOnly": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of last user update",
            "readOnly": true
          },
          "is_protected": {
            "type": "boolean",
            "description": "Indicates if this user is protected from deletion or certain updates",
            "default": false
          }
        }
      },
...
      "Token": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string",
            "default": "bearer"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}